v0.1: Extremely rough proof-of-concept
A proof-of-concept of the RTA minus Time Bonuses plugin using a
hardcoded (for now) lookup table of Sonic 3 time bonuses.

Use like the Manual Game Timer plugin, but make sure your timing method
is set to Real Time instead of Game Time. Enter the in-game time for
each level you complete; the plugin will split, then look up the amount
of time to pause for the entered time and pause automatically for that
length of time.

If you don't need any pause, just hit RETURN for a normal split.

This release is likely to be extremely buggy, and is probably not
suitable for timing. Cleanup will be performed in later releases.