Major timing changes + fix split undoing
We no longer pause the entire LiveSplit timer; instead, we leave the
Real Time alone and pause only the Game Time. This is a much, much
better way of doing things than before as it allows runners to display
both the pure RTA time and the RTA-TB time for each split as well as
have two timers showing at once if desired.

However, this does mean that if you got new Best Segment times (gold
splits) with the old code, these times need to now be moved from the
"Real Time" tab of the splits dialog to the "Game Time" tab as Real Time
is now reverting back to pure RTA time. My apologies for the
inconvenience! If it's any consolation, there shouldn't be any more
major timing changes after this.

This commit also fixes a few other things:

* Timing is now much more accurate. The old code used a Timer control to
  determine when to resume the timer, which was a bad idea but good
  enough for the proof-of-concept. We now use LiveSplit's internal
  timing functionality, which allows us to take advantage of automatic
  NTP syncing and clock drift correction. This fixes issue #2.

* When undoing a split, the component now correctly restores the old
  RTA-TB value, even if you do so when the game timer is paused. This
  fixes issue #1.