Remove useless setting; prepare for release
The "Game Time Input" setting did nothing; it was a vestige from the
Manual Game Time component that RTA-TB is based on. Remove it and
replace with explanatory text explaining how to use the component.

This commit also changes the component name/description to better
describe it, and updates the version number in anticipation of a 0.5
release.