Fix undoing regular splits
If you were to make a split using the regular split key (bypassing the
component) and then undo it, the plugin would not correctly restore the
time because the "last split time" (needed for undoing the split) was
not saved when doing a regular split.

This commit moves the saving of the last split to the OnSplit event,
which is called whenever any split is done, whether by the component or
not.

This fixes issue #6.