Implement adjustable settings!
The plugin now fully supports changeable settings! The settings that can
be adjusted are:

* Input Method: Select between entering in-game time or entering the
                points bonus directly.

* Countdown Speed: Change the "points per frame" and "frames per second"
                   values to adjust the pause time accordingly.

* IGT Lookup Table: The most important setting, this is the lookup table
                    used to convert IGT input to points input so that
                    the plugin pauses for the correct length of time.

With these changes, the plugin can now handle games other than Sonic 3!
I'm not making a public release yet as it still needs to be tested, and
I believe this is the single biggest change yet so there are plenty of
opportunities for bugs and glitches to arise.

This fixes issue #4, and kinda-sorta fixes issue #5 but not really; the
points input is very bare-bones at the moment (it's still just a
textbox) and really needs better methods to input a score easily.