Change background colour of text box on error
Previously, if an error occurred while trying to parse your input during
a run, you'd only know because the text box wasn't cleared.  Now, the
background colour will change too, giving a more visible hint that
something went wrong.

(There are certain circumstances where this will only occur after a
split has happened; for example, if the "Points per frame" setting is
set to 0, then even a valid input will cause the textbox to be
highlighted as it tried to work out how long to pause for, which in this
could would happen after the split itself.)