Fix resizing of time window
Minor fix to make resizing the in-game time entry window do the same as
it did for Manual Game Time - resizing the text box and centering it.

Note that this does not add any code to save the size of the window
between runs. (The position of the window is still saved, as always.)