Release 0.7.1: Set default FPS value to 59.9228Hz
It's been pointed out that the default FPS value of the plugin, 59.94Hz,
isn't correct for playing Sonic 3 on the Genesis. This is a quick
release that fixes this by adding a new radio button for 59.9228Hz to
the settings window and making that the default for new instances of the
component (or updates from an older version with no configurable
settings). This release has no other changes.